requirements.txt
fastapi==0.129.0
uvicorn[standard]==0.40.0
Pillow==9.0.1
pydantic==2.12.5
requests==2.32.5
python-multipart==0.0.22
python-dotenv==1.0.1
